frontend/streamlit/requirements.txt
# --- Core dependencies ---
streamlit==1.39.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
streamlit-chat==0.1.1

# --- Development & quality of life ---
black==24.8.0
flake8==7.1.1
pytest==8.3.3
watchdog==4.0.1
rich==13.9.2

# --- Optional (useful for type hints & safety) ---
typing-extensions==4.12.2
extra-streamlit-components==0.1.71



